requirements.txt
asgiref==3.8.1
Django==5.1.2
gunicorn==23.0.0
packaging==24.1
python-dotenv==1.0.1
sqlparse==0.5.1
typing_extensions==4.12.2